requirements.txt
pyrogram==2.0.106
tgcrypto
psutil
python-dotenv
